Moved code from vc and vole stream files back to vc and vole
