+++ meson.build
@@ -449,8 +449,6 @@ faest_sources = files(
   'vc.c',
   'vole.c',
   'vbb.c',
-  'vole_stream.c',
-  'vc_stream.c',
 )
 
 # header files to install
